projects/auth_management/requirements.txt
# Зависимости для автоматизации тестирования на Playwright
# Основные зависимости
playwright>=1.52.0
pytest>=8.4.0
pytest-playwright>=0.7.0
requests>=2.32.0

# Вспомогательные библиотеки
python-dotenv>=1.1.0
colorama>=0.4.6
typing_extensions>=4.14.0

# Отчетность
allure-pytest>=2.14.0
allure-python-commons>=2.14.0

# Утилиты
python-slugify>=8.0.0
Pygments>=2.19.0

# --- ДОБАВЛЕНО ДЛЯ КОРРЕКТНОЙ РАБОТЫ ПРОЕКТА ---
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
bcrypt>=4.1.2
# sqlite3 входит в стандартную библиотеку Python, отдельная установка не требуется
